Removed scars
I removed scars from the game in order to make the game more peaceful.  The code is still there in case I change my mind later, but as of right now, you should not be able to scar people.

When you buy an upgrade that would have a scar, the scar is still on the weapon so that if I reenable it, you will still be able to use past weapons.  Also, past scars remain.